tests/wpt/web-platform-tests/svg/painting/parsing/paint-order-valid.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml"
     width="800px" height="800px">
  <title>SVG Painting: parsing paint-order with valid values</title>
  <metadata>
    <h:link rel="help" href="https://svgwg.org/svg2-draft/painting.html#PaintOrderProperty"/>
    <h:meta name="assert" content="paint-order supports the full grammar 'normal | [ fill || stroke || markers ]'."/>
    <h:meta name="assert" content="paint-order uses the shortest serialization."/>
  </metadata>
  <g id="target"></g>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <h:script src="/css/support/parsing-testcommon.js"/>
  <script><![CDATA[

test_valid_value("paint-order", "normal");

test_valid_value("paint-order", "fill");
test_valid_value("paint-order", "stroke");
test_valid_value("paint-order", "markers");

test_valid_value("paint-order", "fill stroke", "fill");
test_valid_value("paint-order", "fill markers");
test_valid_value("paint-order", "stroke fill", "stroke");
test_valid_value("paint-order", "stroke markers");
test_valid_value("paint-order", "markers fill", "markers");
test_valid_value("paint-order", "markers stroke");

test_valid_value("paint-order", "fill stroke markers", "fill");
test_valid_value("paint-order", "fill markers stroke", "fill markers");
test_valid_value("paint-order", "stroke fill markers", "stroke");
test_valid_value("paint-order", "stroke markers fill", "stroke markers");
test_valid_value("paint-order", "markers fill stroke", "markers");
test_valid_value("paint-order", "markers stroke fill", "markers stroke");

  ]]></script>
</svg>
